+++ src/keen/__main__.py
@@ -1,0 +1,8 @@
+# -*- coding: utf-8 -*-
+
+"""KEEN main."""
+
+from keen.cli import main
+
+if __name__ == '__main__':
+    main()
