+++ src/keen/cli.py
@@ -1,12 +1,13 @@
+# -*- coding: utf-8 -*-
 import json
 import os
 import pickle
 import sys
 import time
 
-# import matplotlib
-# import matplotlib.pyplot as plt
-# matplotlib.matplotlib_fname()
+import click
+
+from keen.constants import *
 
 w_dir = os.path.dirname(os.getcwd())
 sys.path.append(w_dir)
@@ -479,7 +480,6 @@ def select_training_model_params(model_id):
         if selected_model == TRANS_E:
             kg_model_params[NORM_FOR_NORMALIZATION_OF_ENTITIES] = select_norm(ENTITIES_NORMALIZATION_PRINT_MSG)
 
-
         if selected_model == TRANS_H:
             kg_model_params[WEIGHT_SOFT_CONSTRAINT_TRANS_H] = select_float_value(
                 WEIGHT_SOFT_CONSTRAINT_TRANS_H_PRINT_MSG, WEIGHT_SOFT_CONSTRAINT_TRANS_H_PROMPT_MSG,
@@ -618,7 +618,9 @@ def start_cli():
     return config
 
 
+@click.command()
 def main():
+    """KEEN: A software for training and evaluating knowledge graph embeddings."""
     config = start_cli()
 
     current_time = time.strftime("%H:%M:%S")
@@ -659,7 +661,3 @@ def main():
     out_path = os.path.join(output_direc, 'losses.txt')
     with open(out_path, 'w') as handle:
         handle.write(json.dumps(loss_per_epoch))
-
-
-if __name__ == '__main__':
-    main()
