+++ src/keen/kg_embeddings_model/trans_h.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
+
 import numpy as np
 import torch
 import torch.autograd
 import torch.nn as nn
 
-from keen.utilities import EMBEDDING_DIM, MARGIN_LOSS, NUM_ENTITIES, NUM_RELATIONS, TRANS_H, \
-    WEIGHT_SOFT_CONSTRAINT_TRANS_H, SCORING_FUNCTION_NORM, PREFERRED_DEVICE, GPU, CPU
+from keen.constants import *
 
 'Based on https://github.com/thunlp/OpenKE/blob/OpenKE-PyTorch/models/TransH.py'
