+++ src/keen/utilities/initialization_utils/module_initialization_utils.py
@@ -1,11 +1,12 @@
 # -*- coding: utf-8 -*-
+
+from keen.constants import *
 from keen.kg_embeddings_model.conv_e import ConvE
 from keen.kg_embeddings_model.rot_e import RotE
 from keen.kg_embeddings_model.trans_d import TransD
 from keen.kg_embeddings_model.trans_e import TransE
 from keen.kg_embeddings_model.trans_h import TransH
 from keen.kg_embeddings_model.trans_r import TransR
-from keen.utilities import TRANS_E, TRANS_H, KG_EMBEDDING_MODEL, TRANS_D, TRANS_R, CONV_E, ROT_E
 
 
 def get_kg_embedding_model(config):
@@ -28,5 +29,3 @@ def get_kg_embedding_model(config):
         return ConvE(config=config)
     if model_name == ROT_E:
         return RotE(config=config)
-
-
